💎 #90 Tree v2
Added `GenTree` to be the new underlying problem-space explorer. It has the ability to track "complexity" of a node, so we are able to figure out what is the smallest failure between multiple runs of a property.

It also has a much smarter `expand` function (`expand` is now built into `merge`). It is able to reduce the size of a given forest before it shrinks individual trees. e.g. [1,1] goes to [1] before it goes to [0,1]. This means that shrinking arrays will scale O(log n). The previous implementation did not do this as it is very difficult to transpose the prior shrinks onto the array of shrunken length, e.g. after going from [1,1] to [1], it was hard to then go to [0].