💎 #160 GenTree.unfold has repetitition protection built-in
A shrink function is not aware about the location of the generated shrinks in the tree. It's common for a shrink to be repeated. For example, if we are shrinking to 0, 0 will be the first shrink attempted at each level. GenTree.unfold is one of the basic primitives of other gens - it's the mechanism by which integer generation builds it's GenTree. Add a hook for an equality function that ensures a node has not being encountered before, as the tree is explored.

Performance of this is not scalable yet. It's O(n) with number of shrink iterations traversed. We should be able to crank it a bit if we use a hashing function instead of an equality one (and track encountered values with a Set<string> of hashes).