💎 #160 Build reptition protection into GenTree
Most of the shrink functions work by starting at the most trivial values (e.g. 0, or an empty array), and edging towards the current value. Shrink functions are applied recursively in order to build up trees (GenTrees). Shrink functions are unaware of the level of the tree in which they are operating. Therefore, unless the current value is the trivial value, they will generate the trivial value at each level of the tree. This means that the trivial values are commonly repeated in a GenTree, which has the implication of properties being unnecessarily repeated with an empty array, 0, or other trivial values (also less trivial values are repeated, albiet less frequently).

***We can remedy this by keeping track of what values we have traversed in the tree, and omitting ones we have seen before***

A full-blown equality test would be expensive. Also, we might be comparing things that aren't naturally equatable in JavaScript. An easy example would be generating functions. Two functions are equal if they generate the same output forall inputs. However, all of the most interesting generators (like a generator for funcs) are composed from primitive building blocks. An example of a primitive building block would be the generator for integers. A generator for integers is super-close to the raw RNG aspect of the library, which is where all random values have an origin. These primitive values are also easily equatable, by way of comparing integers.

This commit adds an `id` property to each node in the GenTree. A node's id is originally seeded during a GenTree.unfold, and is customisable by providing an `identify` function. It's not much of a tax to require such a function, as it is not common to roll a primitive generator, the existing primitive generators exist in the library. The identify function for an integer generator simply stringifies the int.

Ids are composed in higher-order generators in a way that is isomorphic to the composition of the node's value. For example, in the array generator, the ids of the nodes are joined together to create a new id. The array of [0,1] gets the id ['0', '1']

For mapping operations, ids are persisted. For example, for the alpha-char generator `Gen.integer().between(0, 25).map((x) => String.fromCharCode(x + 65))`, which generates charactors between 'A' and 'Z', the id of 'A' will be '0', and the id of 'Z' will be 25. It's ok that the id doesn't have the type of the generator encoded into it, as it is only used to identify repeated values within a single generator (i.e. it only discriminates values of the same type).

The set used to track ids is a custom implementation, and is aware of the structure of ids. It builds up a tree based on the components of an id, so we can tell really quickly if we've seen an id by reading that id part-by-part and traversing that tree. For example, given the id ['0, '1'], if we haven't seen any ids starting with '0', we know we haven't seen this id yet.